requirements.txt
pandas
sentence-transformers
pinecone
google-generativeai
grpcio
protobuf
streamlit
fastapi
uvicorn
